vridge_back/requirements.txt
django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
dj-database-url==2.1.0
gunicorn==21.2.0
whitenoise==6.6.0
django-environ==0.11.2
requests==2.31.0
celery==5.3.4
redis==5.0.1
django-redis==5.4.0
ffmpeg-python==0.2.0
pillow==10.1.0